requirements.txt
google-adk
python-dotenv
requests
google-cloud-bigquery
google-generativeai